refactor(Header): add JSDoc, improve structure, optimize event listeners, removes inline comments, rename file
