+++ src/components/Header/Header.jsx
@@ -7,6 +7,10 @@ import ProfileDropdown from './ProfileDropdown';
 import MobileMenu from './MobileMenu';
 import LinkButton from '../Button/LinkButton';
 
+/**
+ * Header component containing navigation, profile dropdown, and mobile menu.
+ * @returns {JSX.Element} - Rendered Header component.
+ */
 export default function Header() {
   const location = useLocation();
   const isHome = location.pathname === '/';
@@ -16,38 +20,32 @@ export default function Header() {
   const [menuOpen, setMenuOpen] = useState(false);
 
   useEffect(() => {
-    const handleScroll = () => {
-      setScrolled(window.scrollY > 20);
-    };
-
+    const handleScroll = () => setScrolled(window.scrollY > 20);
     window.addEventListener('scroll', handleScroll);
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
 
   return (
     <header
-      className={`${isHome ? 'fixed' : 'sticky'} left-0 top-0 z-50 w-full transition-all duration-300 ${
+      className={`${
+        isHome ? 'fixed' : 'sticky'
+      } left-0 top-0 z-50 w-full transition-all duration-300 ${
         isHome && !scrolled ? 'bg-transparent text-white' : 'bg-bg-primary text-white'
       }`}
     >
       <div className="flex items-center justify-between px-5 py-4">
-        {/* Logo */}
-        <div className="logo font-logo text-4xl">
-          <Link to="/">Holidaze</Link>
-        </div>
+        <Link to="/" className="logo font-logo text-4xl">
+          Holidaze
+        </Link>
 
-        {/* Desktop Navigation */}
-        <div className="hidden items-center space-x-5 md:flex">
-          <nav className="flex space-x-5 text-lg">
-            <div className="transition-transform duration-300 hover:scale-105 hover:text-bg-highlight">
-              <Link to="/">Home</Link>
-            </div>
-            <div className="transition-transform duration-300 hover:scale-105 hover:text-bg-highlight">
-              <Link to="/venues">Venues</Link>
-            </div>
-          </nav>
+        <nav className="hidden items-center space-x-5 md:flex">
+          <Link to="/" className="text-lg transition-transform duration-300 hover:scale-105 hover:text-bg-highlight">
+            Home
+          </Link>
+          <Link to="/venues" className="text-lg transition-transform duration-300 hover:scale-105 hover:text-bg-highlight">
+            Venues
+          </Link>
 
-          {/* Profile Dropdown or Login/Register */}
           {user ? (
             <ProfileDropdown
               user={user}
@@ -61,21 +59,14 @@ export default function Header() {
               <LinkButton to="/register" label="Register" variant="highlight" />
             </div>
           )}
-        </div>
+        </nav>
 
-        {/* Mobile Menu Button */}
         <button onClick={() => setMenuOpen(!menuOpen)} className="text-3xl md:hidden">
           <MdMenu />
         </button>
       </div>
 
-      {/* Mobile Menu Component */}
-      <MobileMenu
-        user={user}
-        menuOpen={menuOpen}
-        setMenuOpen={setMenuOpen}
-        logoutUser={logoutUser}
-      />
+      <MobileMenu user={user} menuOpen={menuOpen} setMenuOpen={setMenuOpen} logoutUser={logoutUser} />
     </header>
   );
 }
