+++ src/components/Header/MobileMenu.jsx
@@ -1,20 +1,27 @@
 import { Link } from 'react-router-dom';
 import { MdClose } from 'react-icons/md';
 
+/**
+ * MobileMenu component for navigation on smaller screens.
+ * @param {Object} props - Component props.
+ * @param {Object|null} props.user - Authenticated user data.
+ * @param {boolean} props.menuOpen - Whether the menu is open.
+ * @param {Function} props.setMenuOpen - Function to toggle menu state.
+ * @param {Function} props.logoutUser - Function to log out the user.
+ * @returns {JSX.Element} - Rendered MobileMenu component.
+ */
 export default function MobileMenu({ user, menuOpen, setMenuOpen, logoutUser }) {
   return (
     <div
-      className={`fixed left-0 top-0 z-40 h-full w-full transform bg-black bg-opacity-50 transition-transform ${
+      className={`fixed left-0 top-0 z-40 h-full w-full bg-black bg-opacity-50 transition-transform md:hidden ${
         menuOpen ? 'translate-x-0 opacity-100' : 'translate-x-full opacity-0'
-      } md:hidden`}
+      }`}
     >
       <div className="absolute right-0 top-0 h-full w-64 bg-white p-6 text-black shadow-lg">
-        {/* Close Button */}
         <button onClick={() => setMenuOpen(false)} className="absolute right-4 top-4 text-2xl">
           <MdClose />
         </button>
 
-        {/* Mobile Navigation */}
         <nav className="mt-10 flex flex-col space-y-4 text-lg">
           <Link to="/" className="hover:text-bg-highlight" onClick={() => setMenuOpen(false)}>
             Home
@@ -25,47 +32,25 @@ export default function MobileMenu({ user, menuOpen, setMenuOpen, logoutUser }) 
 
           {user ? (
             <>
-              <Link
-                to="/profile"
-                className="hover:text-bg-highlight"
-                onClick={() => setMenuOpen(false)}
-              >
+              <Link to="/profile" className="hover:text-bg-highlight" onClick={() => setMenuOpen(false)}>
                 Profile
               </Link>
               {user.venueManager && (
-                <Link
-                  to="/venues/new"
-                  className="hover:text-bg-highlight"
-                  onClick={() => setMenuOpen(false)}
-                >
+                <Link to="/venues/new" className="hover:text-bg-highlight" onClick={() => setMenuOpen(false)}>
                   Add Venue
                 </Link>
               )}
               <div className="border-t border-gray-300"></div>
-              <button
-                onClick={() => {
-                  logoutUser();
-                  setMenuOpen(false);
-                }}
-                className="rounded-full py-2 text-left text-red-500"
-              >
+              <button onClick={() => { logoutUser(); setMenuOpen(false); }} className="rounded-full py-2 text-left text-red-500 hover:text-red-700">
                 Logout
               </button>
             </>
           ) : (
             <>
-              <Link
-                to="/login"
-                className="hover:text-bg-highlight"
-                onClick={() => setMenuOpen(false)}
-              >
+              <Link to="/login" className="hover:text-bg-highlight" onClick={() => setMenuOpen(false)}>
                 Login
               </Link>
-              <Link
-                to="/register"
-                className="hover:text-bg-highlight"
-                onClick={() => setMenuOpen(false)}
-              >
+              <Link to="/register" className="hover:text-bg-highlight" onClick={() => setMenuOpen(false)}>
                 Register
               </Link>
             </>
