+++ src/components/Header/ProfileDropdown.jsx
@@ -1,6 +1,15 @@
 import { Link } from 'react-router-dom';
 import { useRef, useEffect } from 'react';
 
+/**
+ * ProfileDropdown component for user account actions.
+ * @param {Object} props - Component props.
+ * @param {Object} props.user - User data.
+ * @param {boolean} props.dropdownOpen - Whether dropdown is open.
+ * @param {Function} props.setDropdownOpen - Function to toggle dropdown state.
+ * @param {Function} props.logoutUser - Function to log out the user.
+ * @returns {JSX.Element} - Rendered ProfileDropdown component.
+ */
 export default function ProfileDropdown({ user, dropdownOpen, setDropdownOpen, logoutUser }) {
   const dropdownRef = useRef(null);
 
@@ -11,12 +20,7 @@ export default function ProfileDropdown({ user, dropdownOpen, setDropdownOpen, l
       }
     };
 
-    if (dropdownOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-    } else {
-      document.removeEventListener('mousedown', handleClickOutside);
-    }
-
+    document.addEventListener('mousedown', handleClickOutside);
     return () => document.removeEventListener('mousedown', handleClickOutside);
   }, [dropdownOpen]);
 
@@ -27,44 +31,25 @@ export default function ProfileDropdown({ user, dropdownOpen, setDropdownOpen, l
         className="flex items-center space-x-2 rounded-md bg-transparent py-2 transition-transform duration-300 hover:scale-105 hover:text-bg-highlight"
       >
         <p>Profile</p>
-        <img
-          src={user.avatar?.url || 'https://via.placeholder.com/40'}
-          alt="Profile"
-          className="h-8 w-8 rounded-full border border-white"
-        />
+        <img src={user.avatar?.url || 'https://via.placeholder.com/40'} alt="Profile" className="h-8 w-8 rounded-full border border-white" />
       </button>
 
-      {/* Dropdown Menu */}
       <div
-        className={`absolute right-0 z-50 mt-2 w-48 transform overflow-hidden rounded-md border bg-white text-black shadow-lg transition-transform ${
+        className={`absolute right-0 z-50 mt-2 w-48 overflow-hidden rounded-md border bg-white text-black shadow-lg transition-transform origin-top ${
           dropdownOpen ? 'scale-y-100 opacity-100' : 'scale-y-0 opacity-0'
-        } origin-top`}
+        }`}
       >
-        <Link
-          to="/profile"
-          className="block px-4 py-2 hover:bg-gray-100"
-          onClick={() => setDropdownOpen(false)}
-        >
+        <Link to="/profile" className="block px-4 py-2 hover:bg-gray-100" onClick={() => setDropdownOpen(false)}>
           View Profile
         </Link>
 
         {user.venueManager && (
-          <Link
-            to="/venues/new"
-            className="block px-4 py-2 hover:bg-gray-100"
-            onClick={() => setDropdownOpen(false)}
-          >
+          <Link to="/venues/new" className="block px-4 py-2 hover:bg-gray-100" onClick={() => setDropdownOpen(false)}>
             Add Venue
           </Link>
         )}
         <div className="border-t border-gray-300"></div>
-        <button
-          onClick={() => {
-            logoutUser();
-            setDropdownOpen(false);
-          }}
-          className="block w-full px-4 py-2 text-left text-red-500 hover:bg-gray-100"
-        >
+        <button onClick={() => { logoutUser(); setDropdownOpen(false); }} className="block w-full px-4 py-2 text-left text-red-500 hover:bg-gray-100">
           Logout
         </button>
       </div>
