+++ src/components/Layout/index.jsx
@@ -1,6 +1,6 @@
 import { Outlet, useLocation } from 'react-router-dom';
 import Footer from '../Footer/Footer';
-import Header from '../Header';
+import Header from '../Header/Header';
 
 export default function Layout() {
   const location = useLocation();
